requirements.txt
streamlit
numpy
matplotlib
trimesh
numpy-stl
